Move error enums and impl in lib.rs to error.rs
